@@ -8,47 +8,45 @@ version: 2.7.02+23
 environment:
   # sdk: ">=2.12.0 <3.0.0"
   # sdk: ">=2.18.5 <3.0.0"
-  sdk: ">=3.0.0 <4.0.0"
-
-dependency_overrides:
-  intl: ^0.18.0
+  # sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.4.3 <4.0.0"
 
 dependencies:
   audioplayers: ^3.0.0
   auto_size_text: ^3.0.0
-  awesome_notifications: ^0.7.4+1
+  awesome_notifications: ^0.9.3
   bloc: ^8.1.3
   cupertino_icons: ^1.0.4
-  day_night_time_picker: ^1.0.5
+  day_night_time_picker: ^1.3.1
   enum_to_string: ^2.0.1
   equatable: ^2.0.5
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.4
-  flutter_colorpicker: ^1.0.3
-  flutter_local_notifications: ^13.0.0
+  flutter_bloc: ^8.1.6
+  flutter_colorpicker: ^1.1.0
+  flutter_local_notifications: ^17.2.1
   flutter_localizations:
     sdk: flutter
-  flutter_native_splash: ^2.1.6
-  flutter_slidable: ^2.0.0
-  flutter_zoom_drawer: ^3.0.1
-  fluttertoast: ^8.0.8
-  font_awesome_flutter: ^10.1.0
-  get: ^4.6.5
-  get_storage: ^2.0.3
-  intl: ^0.18.0
-  material_design_icons_flutter: ^6.0.7096
-  path: ^1.8.2
-  path_provider: ^2.0.8
-  rxdart: ^0.27.3
+  flutter_native_splash: ^2.4.1
+  flutter_slidable: ^3.1.0
+  flutter_zoom_drawer: ^3.2.0
+  fluttertoast: ^8.2.6
+  font_awesome_flutter: ^10.7.0
+  get: ^4.6.6
+  get_storage: ^2.1.1
+  intl: ^0.19.0
+  material_design_icons_flutter: ^7.0.7296
+  path: ^1.9.0
+  path_provider: ^2.1.3
+  rxdart: ^0.28.0
   share: ^2.0.4
-  shared_preferences: ^2.0.12
+  shared_preferences: ^2.2.3
   sleek_circular_slider: ^2.0.1
-  sqflite: ^2.0.2
+  sqflite: ^2.3.3
   timelines: ^0.1.0
-  url_launcher: ^6.1.7
-  vibration: ^1.7.4-nullsafety.0
-  wakelock: ^0.6.1+1
+  url_launcher: ^6.3.0
+  vibration: ^1.7.7
+  wakelock: ^0.6.2
 
 dev_dependencies:
   flutter_lints: ^2.0.1
