@@ -140,12 +140,12 @@ class AddAlarmDialogState extends State<AddAlarmDialog> {
               Navigator.of(context).push(
                 showPicker(
                   context: context,
-                  value: _time,
+                  value: Time(hour: _time.hour, minute: _time.minute),
                   onChange: onTimeChanged,
                   iosStylePicker: true,
                   // Optional onChange to receive value as DateTime
                   onChangeDateTime: (DateTime dateTime) {},
-                ),
+                ) as Route,
               );
             },
           ),
