@@ -68,7 +68,7 @@ class FavouriteZikr extends StatelessWidget {
                                     Expanded(
                                       child: IconButton(
                                         splashRadius: 20,
-                                        icon: const Icon(MdiIcons.camera),
+                                        icon: Icon(MdiIcons.camera),
                                         onPressed: () {
                                           transitionAnimation.circleReval(
                                             context: Get.context!,
