@@ -14,15 +14,17 @@ class ScreenAppBar extends StatelessWidget {
   Widget build(BuildContext context) {
     return SliverAppBar(
       leading: !controller.isSearching
-          ? Image.asset(
-              'assets/images/app_icon.png',
-              width: 20,
-              height: 20,
+          ? Padding(
+              padding: const EdgeInsets.all(7),
+              child: Image.asset(
+                'assets/images/app_icon.png',
+                fit: BoxFit.cover,
+              ),
             )
           : IconButton(
               splashRadius: 20,
               padding: EdgeInsets.zero,
-              icon: const Icon(MdiIcons.close),
+              icon: Icon(MdiIcons.close),
               onPressed: () {
                 controller.isSearching = false;
                 controller.searchedTitle = controller.allTitle;
@@ -50,7 +52,7 @@ class ScreenAppBar extends StatelessWidget {
                   right: 15,
                 ),
                 suffix: IconButton(
-                  icon: const Icon(MdiIcons.eraser),
+                  icon: Icon(MdiIcons.eraser),
                   onPressed: () {
                     controller.searchController.clear();
                     controller.searchZikr();
