@@ -4,7 +4,7 @@ import 'package:get/get.dart';
 import 'package:hisnelmoslem/src/core/shared/widgets/loading.dart';
 import 'package:hisnelmoslem/src/core/values/app_dashboard.dart';
 import 'package:hisnelmoslem/src/features/home/presentation/components/screen_appbar.dart';
-import 'package:hisnelmoslem/src/features/home/presentation/components/screen_menu.dart';
+import 'package:hisnelmoslem/src/features/home/presentation/components/side_menu/side_menu.dart';
 import 'package:hisnelmoslem/src/features/home/presentation/controller/dashboard_controller.dart';
 import 'package:hisnelmoslem/src/features/settings/presentation/components/rearrange_dashboard/rearrange_dashboard_page_controller.dart';
 import 'package:intl/intl.dart';
@@ -20,22 +20,24 @@ class AzkarDashboard extends StatelessWidget {
       init: DashboardController(),
       builder: (controller) => controller.isLoading
           ? const Loading()
-          : ZoomDrawer(
-              isRtl: Bidi.isRtlLanguage(Get.locale!.languageCode),
-              controller: controller.zoomDrawerController,
-              menuBackgroundColor: Theme.of(context).scaffoldBackgroundColor,
-              menuScreen: ScreenMenu(
-                controller: controller,
+          : Scaffold(
+              body: ZoomDrawer(
+                isRtl: Bidi.isRtlLanguage(Get.locale!.languageCode),
+                controller: controller.zoomDrawerController,
+                menuBackgroundColor: Theme.of(context).scaffoldBackgroundColor,
+                menuScreen: SideMenu(
+                  controller: controller,
+                ),
+                mainScreen: MainScreen(
+                  controller: controller,
+                ),
+                borderRadius: 24.0,
+                showShadow: true,
+                angle: 0.0,
+                drawerShadowsBackgroundColor:
+                    Theme.of(context).colorScheme.primary,
+                slideWidth: 270,
               ),
-              mainScreen: MainScreen(
-                controller: controller,
-              ),
-              borderRadius: 24.0,
-              showShadow: true,
-              angle: 0.0,
-              drawerShadowsBackgroundColor:
-                  Theme.of(context).colorScheme.primary,
-              slideWidth: 270,
             ),
     );
   }
