@@ -30,7 +30,7 @@ class ZikrViewerPageModeAppBar extends StatelessWidget {
                 Expanded(
                   child: IconButton(
                     splashRadius: 20,
-                    icon: const Icon(MdiIcons.comment),
+                    icon: Icon(MdiIcons.comment),
                     onPressed: () {
                       showCommentaryDialog(
                         context: Get.context!,
@@ -43,7 +43,7 @@ class ZikrViewerPageModeAppBar extends StatelessWidget {
                 Expanded(
                   child: IconButton(
                     splashRadius: 20,
-                    icon: const Icon(MdiIcons.camera),
+                    icon: Icon(MdiIcons.camera),
                     onPressed: () {
                       transitionAnimation.circleReval(
                         context: Get.context!,
